+++ Sources/BraintreeSEPADirectDebit/BTSEPADirectDebitNonce.swift
@@ -23,9 +23,9 @@ import BraintreeCoreSwift
     // NEXT_MAJOR_VERSION consider making init non-optional (especially internal ones) once we convert to Swift.
     init?(json: BTJSON) {
         guard let nonce = json["nonce"].asString() else { return nil }
-        
-        self.ibanLastFour = json["details"]["last4"].asString()
-        self.customerID = json["details"]["customerId"].asString()
+
+        self.ibanLastFour = json["details"]["ibanLastChars"].asString()
+        self.customerID = json["details"]["merchantOrPartnerCustomerId"].asString()
         self.mandateType = BTSEPADirectDebitMandateType.getMandateType(from: json["details"]["mandateType"].asString())
 
         super.init(nonce: nonce, type: "SEPADebit")