+++ UnitTests/BraintreeCoreTests/BTClientToken_Tests.swift
@@ -11,7 +11,7 @@ class BTClientToken_Tests: XCTestCase {
             XCTAssertNil(clientToken)
         } catch let error as NSError {
             XCTAssertEqual(error.domain, BTClientTokenError.errorDomain)
-            XCTAssertEqual(error.code, BTClientTokenError.unsupportedVersion.rawValue)
+            XCTAssertEqual(error.code, 3)
             XCTAssertEqual(error.localizedDescription, BTClientTokenError.unsupportedVersion.localizedDescription)
         }
     }
@@ -46,8 +46,46 @@ class BTClientToken_Tests: XCTestCase {
             XCTAssertNil(clientToken)
         } catch let error as NSError {
             XCTAssertEqual(error.domain, BTClientTokenError.errorDomain)
-            XCTAssertEqual(error.code, BTClientTokenError.invalidJSON.rawValue)
-            XCTAssertEqual(error.localizedDescription, BTClientTokenError.invalidJSON.localizedDescription)
+            XCTAssertEqual(error.code, 2)
+            XCTAssertEqual(error.localizedDescription, "Invalid client token format. Please ensure your server is generating a valid Braintree ClientToken. Invalid JSON. Expected to find an object at JSON root.")
+        }
+    }
+    
+    func testInitialization_withNilVersion_returnsError() {
+        do {
+            let clientTokenRawJSON = TestClientTokenFactory.token(withVersion: 1, overrides: ["version": nil])
+            let clientToken = try BTClientToken(clientToken: clientTokenRawJSON)
+            XCTAssertNil(clientToken)
+        } catch let error as NSError {
+            XCTAssertEqual(error.domain, BTClientTokenError.errorDomain)
+            XCTAssertEqual(error.code, 2)
+            XCTAssertEqual(error.localizedDescription, "Invalid client token format. Please ensure your server is generating a valid Braintree ClientToken. Invalid version number. Expected to find an integer for key \"version\".")
+        }
+    }
+    
+    func testInitialization_withVersion1Base64Encoding_returnsError() {
+        do {
+            // Force factory to encode for version 2, then override version number for test case
+            let clientTokenRawJSON = TestClientTokenFactory.token(withVersion: 2, overrides: ["version": 1])
+            let clientToken = try BTClientToken(clientToken: clientTokenRawJSON)
+            XCTAssertNil(clientToken)
+        } catch let error as NSError {
+            XCTAssertEqual(error.domain, BTClientTokenError.errorDomain)
+            XCTAssertEqual(error.code, 4)
+            XCTAssertEqual(error.localizedDescription, "Failed to decode client token. UTF8 encoding is required for Client Token version 1.")
+        }
+    }
+    
+    func testInitialization_withVersion2UTF8Encoding_returnsError() {
+        do {
+            // Force factory to encode for version 1, then override version number for test case
+            let clientTokenRawJSON = TestClientTokenFactory.token(withVersion: 1, overrides: ["version": 2])
+            let clientToken = try BTClientToken(clientToken: clientTokenRawJSON)
+            XCTAssertNil(clientToken)
+        } catch let error as NSError {
+            XCTAssertEqual(error.domain, BTClientTokenError.errorDomain)
+            XCTAssertEqual(error.code, 4)
+            XCTAssertEqual(error.localizedDescription, "Failed to decode client token. Base64 encoding is required for Client Token versions 2 & 3.")
         }
     }
 
@@ -60,19 +98,19 @@ class BTClientToken_Tests: XCTestCase {
             XCTAssertNil(clientToken)
         } catch let error as NSError {
             XCTAssertEqual(error.domain, BTClientTokenError.errorDomain)
-            XCTAssertEqual(error.code, BTClientTokenError.invalidConfigURL.rawValue)
+            XCTAssertEqual(error.code, 1)
             XCTAssertEqual(error.localizedDescription, BTClientTokenError.invalidConfigURL.localizedDescription)
         }
     }
 
     func testInitialization_whenAuthorizationFingerprintIsOmitted_returnsError() {
         do {
-            let clientTokenRawJSON = TestClientTokenFactory.token(withVersion: 2, overrides: ["authorizationFingerprint": NSNull()])
+            let clientTokenRawJSON = TestClientTokenFactory.token(withVersion: 2, overrides: ["authorizationFingerprint": nil])
             let clientToken = try BTClientToken(clientToken: clientTokenRawJSON)
             XCTAssertNil(clientToken)
         } catch let error as NSError {
             XCTAssertEqual(error.domain, BTClientTokenError.errorDomain)
-            XCTAssertEqual(error.code, BTClientTokenError.invalidAuthorizationFingerprint.rawValue)
+            XCTAssertEqual(error.code, 0)
             XCTAssertEqual(error.localizedDescription, BTClientTokenError.invalidAuthorizationFingerprint.localizedDescription)
         }
     }
@@ -84,7 +122,7 @@ class BTClientToken_Tests: XCTestCase {
             XCTAssertNil(clientToken)
         } catch let error as NSError {
             XCTAssertEqual(error.domain, BTClientTokenError.errorDomain)
-            XCTAssertEqual(error.code, BTClientTokenError.invalidAuthorizationFingerprint.rawValue)
+            XCTAssertEqual(error.code, 0)
             XCTAssertEqual(error.localizedDescription, BTClientTokenError.invalidAuthorizationFingerprint.localizedDescription)
         }
     }